Add script to run test cases
Calls make and then runs all the tests with your shim
Updated make_help.sh to work with it

+++ make_help.sh
@@ -3,9 +3,14 @@ ls *.c | sed G | grep func >> files.c
 # ^ get all the test files
 FILECOUNT=$(wc -l < files.c)
 # and how many there are
-for ((i=1; i<=FILECOUNT; i++)); do
+for ((i=0; i<=FILECOUNT-1; i++)); do
     FILE=$(awk -v i="$i" 'NR == i' files.c)
     # get names one at a time and compile them
     ($CC $FLAGS $DEBUG "$FILE" -o "$(basename -s .c "$FILE").exe")
 done
-rm files.c # cleanup
+if ! [[ -n "$RUNCALLED" ]]; then
+    # means this was called by make and not run.sh
+    # echo "a script didn't call me"
+    rm files.c # cleanup
+    unset FILECOUNT
+fi
